requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
cycler==0.10.0
enum34==1.1.6
gast==0.2.0
grpcio==1.10.1
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.2
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
Werkzeug==0.14.1
